Import directly from wpcom-request-proxy
